refac: general, broadcast_to and reshape
