+++ ivy/functional/backends/mxnet/data_type.py
@@ -44,3 +44,13 @@ class Finfo:
 def finfo(type: Union[type, str, mx.ndarray.ndarray.NDArray])\
         -> Finfo:
     return Finfo(np.finfo(ivy.dtype_from_str(type)))
+
+def broadcast_to(x, new_shape):
+    x_shape = list(x.shape)
+    num_x_dims = len(x_shape)
+    num_shape_dims = len(new_shape)
+    diff = num_shape_dims - num_x_dims
+    if diff == 0:
+        return mx.nd.broadcast_to(x, new_shape)
+    x = mx.nd.reshape(x, [1]*diff + x_shape)
+    return mx.nd.broadcast_to(x, new_shape)
