+++ ivy/functional/backends/mxnet/manipulation.py
@@ -3,7 +3,8 @@ import mxnet as mx
 import math
 import numpy as np
 from typing import Union, Tuple, Optional, List
-from ivy.functional.backends.mxnet import _flat_array_to_1_dim_array, _handle_flat_arrays_in_out, _handle_flat_arrays_in
+from ivy.functional.backends.mxnet import _flat_array_to_1_dim_array, _handle_flat_arrays_in_out, _handle_flat_arrays_in, _1_dim_array_to_flat_array
+
 
 def flip(x: mx.ndarray.ndarray.NDArray,
          axis: Optional[Union[int, Tuple[int], List[int]]] = None)\
@@ -43,10 +44,14 @@ def squeeze(x, axis=None):
         if axis is None or axis == 0 or axis == -1:
             return x
         raise Exception('tried to squeeze a zero-dimensional input by axis {}'.format(axis))
-    res = _mx.nd.squeeze(x, axis)
+    res = mx.nd.squeeze(x, axis)
     if axis is None:
         return _1_dim_array_to_flat_array(res)
     return res
+
+
+reshape = lambda x, new_shape: x.reshape(new_shape)
+
 
 
 # Extra #
