+++ ivy/functional/backends/mxnet/old/general.py
@@ -135,19 +135,6 @@ def where(condition, x1, x2):
 
 
 
-reshape = lambda x, new_shape: x.reshape(new_shape)
-
-
-def broadcast_to(x, new_shape):
-    x_shape = list(x.shape)
-    num_x_dims = len(x_shape)
-    num_shape_dims = len(new_shape)
-    diff = num_shape_dims - num_x_dims
-    if diff == 0:
-        return _mx.nd.broadcast_to(x, new_shape)
-    x = _mx.nd.reshape(x, [1]*diff + x_shape)
-    return _mx.nd.broadcast_to(x, new_shape)
-
 
 
 # noinspection PyShadowingNames
