+++ ivy/functional/backends/tensorflow/old/general.py
@@ -103,8 +103,6 @@ where = lambda condition, x1, x2: _tf.where(_tf.cast(condition, _tf.bool), x1, x
 
 
 
-reshape = lambda x, newshape: _tf.reshape(x, (newshape,) if isinstance(newshape, int) else newshape)
-broadcast_to = _tf.broadcast_to
 
 
 # noinspection PyShadowingNames
