+++ ivy/functional/backends/torch/manipulation.py
@@ -60,6 +60,13 @@ def stack(xs: List[torch.Tensor], axis: int = 0):
     return torch.stack(xs, axis)
 
 
+
+def reshape(x, newshape: List[int]):
+    if isinstance(newshape, int):
+        newshape = [newshape]
+    return torch.reshape(x, newshape)
+
+
 # Extra #
 # ------#
